requirements.txt
streamlit==1.28.1
openai==1.6.1
python-dotenv==1.0.0
PyPDF2==3.0.1
docx2txt==0.8